fix: 함수 분리 및 콘솔, 주석 삭제
