+++ src/hooks/queries/invitation/useGetInvitationQuery.ts
@@ -1,10 +1,9 @@
 'use client';
-import { createClient } from '@/utils/supabase/client';
+import browserClient from '@/utils/supabase/client';
 import { useQuery } from '@tanstack/react-query';
 import { QUERY_KEYS } from '../queryKeys';
 
 const getExistingInvitation = async () => {
-  const browserClient = createClient();
   const { data: user } = await browserClient.auth.getUser();
 
   if (!user.user) {
